i9: preserve uncommitted work in job-i9-a1-round4 at 2026-09-13 (never pushed)
33 modified tracked file(s) that existed only in this working tree. Preserved as a commit so a disk
event or a mistaken sweep cannot take them. Parent is the directory's HEAD at preservation time.
This branch is LOCAL and must never be pushed.

+++ src/core/runtime/rwmutex.cs
@@ -105,7 +105,7 @@ internal static void runlock(this ж<rwmutex> Ꮡrw) {
 
     {
         var r = Ꮡrw.of(rwmutex.ᏑreaderCount).Add(-1); if (r < 0) {
-            if (r + 1 == 0 || r + 1 == -rwmutexMaxReaders) {
+            if (r + 1 == 0 || r + 1 == (int32)(-rwmutexMaxReaders)) {
                 @throw(runlockOfUnlockedRwmutexˢ);
             }
             // A writer is pending.
